requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
chromadb==1.0.11
sentence-transformers==2.6.1
transformers==4.44.2
torch==2.3.0
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.26.4
requests==2.31.0
langchain==0.3.7
langchain-community==0.3.7
pydantic==2.8.2
typing-extensions==4.12.2
huggingface-hub==0.24.6
tokenizers==0.19.1
safetensors==0.4.3
regex==2024.5.15
tqdm==4.66.5
psutil==5.9.8
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.5
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
urllib3==2.2.2
filelock==3.15.4
fsspec==2024.6.1
packaging==24.1
pyyaml==6.0.2
sympy==1.13
networkx==3.3
mpmath==1.3.0
